+++ stabill/lib/widgets/account_list.dart
@@ -1,5 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
@@ -76,8 +77,8 @@ class _AccountListState extends State<AccountList> {
 
         var accountData = snapshot.data!.docs;
 
-        double totalCurrentBalance = 0;
-        double totalAvailableBalance = 0;
+        int totalCurrentBalance = 0;
+        int totalAvailableBalance = 0;
 
         accountData.forEach((element) {
           Account account = element.data();
